Player info saving
Player information now gets saved and loaded. This includes information like the player's current health, maximum health, gold, spawn, and equipped weapon and armor.

The spawn gets set as the player passes through it, and each of the health, gold, weapon, and armor gets set as it changes.

The UIBar now updates the player's gold, location, and current quest.

+++ sandbox/Assets/Scripts/Attackable/Attackable.cs
@@ -10,8 +10,8 @@ namespace Character2D
         protected SpriteRenderer sr;
         protected Animator anim;
 
-        protected float maxVitality;
-        protected float currentVitality;
+        protected int maxVitality;
+        protected int currentVitality;
 
         public bool isDying;
 
@@ -22,7 +22,7 @@ namespace Character2D
         //used for initialization
         protected void Start()
         {
-            maxVitality = currentVitality = 50f; //TODO: remove
+            maxVitality = currentVitality = 50; //TODO: remove
             isDying = false;
             rb = GetComponent<Rigidbody2D>();
             sr = GetComponent<SpriteRenderer>();
@@ -30,7 +30,7 @@ namespace Character2D
         }
 
         //applies damage to the player
-        public virtual void TakeDamage(GameObject attacker, float damage)
+        public virtual void TakeDamage(GameObject attacker, int damage)
         {
             if (!isDying)
             {
